Settings: move settings tabs to a separate file, move profile tab pane to a tabs_panes folder refs: ZN-57
