+++ app/frontend/javascript/components/settings/tabs_panes/profile/components/profile_settings_form.jsx
@@ -1,4 +1,4 @@
-import FormBase from "../../../common/form_base";
+import FormBase from "../../../../common/form_base";
 import { Form } from "react-bootstrap";
 import { connect } from "react-redux";
